Extract express server instance from probot, change start script
