+++ packages/bot/src/env.ts
@@ -14,6 +14,9 @@ function getEnvInteger(varName: string): number {
   return parsedValue;
 }
 
+export const APP_ID = getEnvInteger("APP_ID");
+export const WEBHOOK_SECRET = getEnvString("WEBHOOK_SECRET");
+export const PRIVATE_KEY = getEnvString("PRIVATE_KEY");
 export const SLACK_SIGNING_SECRET = getEnvString("SLACK_SIGNING_SECRET");
 export const SLACK_CRON_PATTERN = getEnvString("SLACK_CRON_PATTERN");
 export const SLACK_WEBHOOK_URL = getEnvString("SLACK_WEBHOOK_URL");
